+++ charts/finrem-cos/values.preview.template.yaml
@@ -9,6 +9,6 @@ java:
     DOCUMENT_GENERATOR_SERVICE_API_BASEURL : "http://finrem-dgcs-aat.service.core-compute-aat.internal"
     FINREM_NOTIFICATION_SERVICE_BASE_URL : "http://finrem-ns-aat.service.core-compute-aat.internal"
     PAYMENT_SERVICE_API_BASEURL : "http://finrem-ps-aat.service.core-compute-aat.internal"
-    FEATURE_APPROVED_CONSENT_ORDER_NOTIFICATION_LETTER: "false"
-    FEATURE_HWF_SUCCESSFUL_NOTIFICATION_LETTER: "true"
-    FEATURE_ASSIGNED_TO_JUDGE_NOTIFICATION_LETTER: "true"
+    FEATURE_APPROVED_CONSENT_ORDER_NOTIFICATION_LETTER: false
+    FEATURE_HWF_SUCCESSFUL_NOTIFICATION_LETTER: true
+    FEATURE_ASSIGNED_TO_JUDGE_NOTIFICATION_LETTER: true
