+++ src/main/resources/application.properties
@@ -115,7 +115,6 @@ idam.s2s-auth.services-allowed-to-update=${BSP_SERVICES_ALLOWED_TO_UPDATE:bulk_s
 idam.s2s-auth.services-allowed-to-validate=${BSP_SERVICES_ALLOWED_TO_VALIDATE:bulk_scan_processor}
 
 # Features
-feature-toggle.service.api.baseurl = ${FEATURE_TOGGLE_SERVICE_API_BASEURL:http://localhost:4028}
 feature-toggle.approved_consent_order_notification_letter=${FEATURE_APPROVED_CONSENT_ORDER_NOTIFICATION_LETTER:false}
 feature-toggle.hwf_successful_notification_letter=${FEATURE_HWF_SUCCESSFUL_NOTIFICATION_LETTER:false}
 feature-toggle.assigned_to_judge_notification_letter=${FEATURE_ASSIGNED_TO_JUDGE_NOTIFICATION_LETTER:false}
